pyproject.toml
[project]
name = "autolife"
version = "0.1.0"
description = "基于 AutoGLM 的智能助手 - AutoGLM Assistant"
readme = "README.md"
authors = [
    { name = "AutoLife Team", email = "autolife@example.com" }
]
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
keywords = ["assistant", "autoglm", "phone-automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.125.0",
    # AutoGLM 核心依赖
    "openai>=2.9.0",
    "pillow>=12.0.0",
    # "pure-python-adb",  # 暂时注释,需要时手动安装
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn[standard]>=0.38.0",
    # scrcpy 投屏相关（使用命令行工具 + OpenCV）
    "opencv-python>=4.8.0",    # 视频帧处理和 JPEG 编码
    "numpy>=1.24.0",           # 数组处理
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/autolife"
Documentation = "https://github.com/yourusername/autolife/blob/main/docs"
Repository = "https://github.com/yourusername/autolife"
Issues = "https://github.com/yourusername/autolife/issues"

[project.scripts]
autolife = "autolife.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autolife"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
